+++ pkg/goodwe/authentication/authentication.go
@@ -10,9 +10,9 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/AaronSaikovski/gogoodwe/constants"
-	"github.com/AaronSaikovski/gogoodwe/types"
-	"github.com/AaronSaikovski/gogoodwe/utils"
+	"github.com/AaronSaikovski/gogoodwe/pkg/goodwe/constants"
+	"github.com/AaronSaikovski/gogoodwe/pkg/goodwe/utils"
+	"github.com/AaronSaikovski/gogoodwe/pkg/goodwe/types"
 )
 
 // SetHeaders - Set the login headers for the SEMS API login
