+++ pkg/goodwe/authentication/authhelper.go
@@ -9,8 +9,8 @@ import (
 	"errors"
 	"strings"
 
-	"github.com/AaronSaikovski/gogoodwe/constants"
-	"github.com/AaronSaikovski/gogoodwe/types"
+	"github.com/AaronSaikovski/gogoodwe/goodwe/constants"
+	"github.com/AaronSaikovski/gogoodwe/goodwe/types"
 )
 
 // CheckUserLoginInfo - Check user login struct is valid/not null
