+++ pkg/goodwe/constants/constants.go
