+++ pkg/goodwe/powerstationdata/powerstationdatahelper.go
@@ -8,9 +8,8 @@ import (
 	"encoding/json"
 	"strconv"
 
-	"github.com/AaronSaikovski/gogoodwe/types"
-	"github.com/AaronSaikovski/gogoodwe/utils"
-)
+	"github.com/AaronSaikovski/gogoodwe/pkg/goodwe/types"
+	"github.com/AaronSaikovski/gogoodwe/pkg/goodwe/utils"
 
 // DataTokenJSON - Makes a map for the token to be passed to the Data API header and returns a JSON string
 func DataTokenJSON(SemsResponseData *types.SemsResponseData) ([]byte, error) {
