+++ pkg/goodwe/types/semslogincreds.go
