+++ pkg/goodwe/types/semsresponsedata.go
