+++ pkg/goodwe/types/stationresponsedata.go
