+++ pkg/goodwe/utils/errorhandler.go
