+++ pkg/goodwe/utils/paramcheck.go
