+++ pkg/goodwe/utils/responsehandler.go
